move non generic components to featured concept
