+++ components/README.md
@@ -1,0 +1,5 @@
+# Components
+
+This directory contains all the components of the application.
+
+Mainly, we follow a layered architecture for our components, where each layer is responsible for a specific aspect. The sole exception to this is the `featured` directory.
